requirements/common.txt
astroid==1.5.3
Django==1.11.4
django-cors-headers==2.1.0
django-filter==1.0.4
djangorestframework==3.6.4
djangorestframework-oauth==1.1.0
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.7.2
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
pytz==2017.2
six==1.10.0
wrapt==1.10.11
Jinja2==2.9.6
PyYAML==3.12
django-decorator-include==1.3